pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "  "
authors = ["Roman Kazakov <swamper@yandex.ru>"]
packages = [
    {include = 'gendiff'},
]

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[tool.poetry.dependencies]
python = "^3.8"
flake8 = "^4.0.1"
pytest-cov = {version = "^4.0.0", source = "pypi"}
pytest = {version = "^7.2.1", source = "pypi"}
pyyaml = "^6.0"
asciinema = "^2.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
